Update dependencies with Github Action
This Github Action will run every Monday at 3am UTC and use Poetry to update the "main" dependencies. If there are any changes it will use the github command line tool to create a "dependency updates" pull request. Only the "main" dependencies are updated because Airflow is currently being installed by the tests, and we are currently using an Airflow Docker image. We don't want to update the dependencies independently of the Docker image.

Fixes #257