Add a view test
Add a feature test for the vendor listing page in the Vendor Loads
plugin:

- use pytest-mock-resources to create and populate an in memory sqlite
  database with some Vendor fixture data.
- mock Airflow's PostgresHook database connection to return the mocked
  database
- create a test fixture that creates the Airflow Flask application, adds
  our Vendor Loads plugin to it, and returns a web client for talking to
  it.

I put the `test_client` fixture in its own module and loaded it in
`tests.conftest` so that it's available to any other tests that might
need it.

I had hoped to put the database setup and mocking into `tests.mocks` but
since `pytest_mock_resources.create_sqlite_fixture()` is itself a pytest
fixture and you are no longer able to call these directly, I couldn't
figure out a way to do it.