Use libsys_airflow module
This spike adapts the Docker containers to mount the libsys_airflow
module as a volume in /opt/airflow/ and then uses dags_folder and
plugins_folder in the airflow.cfg to point those locations. It also
updates the remaining places that were importing from plugins to import
from libsys_airflow.plugins instead.

The benefit of this approach is that our test code can import from
libsys_airflow just like the code in libsys_airflow that needs to import
from the root module. We could put dag and plugins in the top level, but
I think it's nice to have our airflow code in a named module?