Add Vendor view
Add a vendor view that lists the interfaces for a given vendor. The URL
path is:

    /vendors/{vendor-id}

Initially I had this as:

    /vendors/vendor/{vendor-id}

But it looked kind of weird? We can change it later if it causes
problems with subsequent views?

I moved `tests.client` to `tests.airflow_client` to avoid confusion with
other clients. I also modified the test_airflow_client response so that
it can provide access to a BeautifulSoup object for the response to
allow the tests to operate on the parsed HTML rather than doing straight
up string matching.

Closes #355