Allow black to fix files
With these options in our `pyproject.toml` I couldn't find a way to run
black so that it would fix problems:

        check = true
        diff = true

Usually you can run `black .` on the command line to automatically
reformat all the problems, but it seems like the `pyproject.toml`
settings won't allow that, and black only will "check"?

So I removed the options from `pyproject.toml` and modified the Github
action to use them instead, so it will fail if there is a formatting
probem, and the diff should get displayed -- and we can run black on
the command line to fix things when there are problems.