Add Folio job execution UUID column to VendorFile model
Connects to #508

This commit restores to the VendorFile model a column that was previously removed from the DataLoad model. When the data processor runs and gets the Folio job execution ID, that value is saved to all associated VendorFile instances. This is needed so we can construct links to Folio job executions, e.g., so users can get more information about vendor file loading errors (#508).