Display links to Folio job executions in vendor management dashboard
Fixes #509

NOTE: This commit includes a change to the data import task that makes it possible to associate uploaded files in the vendor management app with files uploaded to Folio, which are always chunked into at least one chunk, which has the effect of adding a `_1` infix to the uploaded filename just before the file extension. So, uploading `data.mrc` to the vendor management app winds up loading `data_1.mrc` in Folio-land. To successfully associate these infixed files with the filenames in the `VendorFile` table, this commit maps over the `batch_filenames` in the data import task, dropping the `_n` infix and uniq'ing the list of filenames, which allows the filenames from Folio-land to match those in the vendor_loads database.