Set status
Add the ability for the user to set the status to "loaded" for
VendorFiles that have not already been loaded or purged.

The select box only allows transitioning from the current status to
"loaded". Once it has been marked loaded (or it has been purged) the
select box does not display and the status view is displayed read-only.

Resolves #497