Change logic that determines if email is enabled
Fixes #645

This commit fixes a bug manifesting in local development where the app tries to send email even though it shouldn't. Why? Because the prior implementation was based on whether the `smtp.smtp_host` Airflow configuration option was set, and that option defaults to being set to `localhost`. Instead, determine that email should not be set if the option is set to its default value.